+++ CHANGELOG.md
@@ -7,6 +7,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [8.7.1] - 2026-08-07
+
 ### Fixed
 
 - Git resource providers now rebuild the local cache from scratch when it exceeds `cache_limit`,
@@ -1398,7 +1400,8 @@ See the Python Migration Guide for details on how to update your YAML configurat
 - Nearly all command line options removed in favor of YAML equivalents.
 - Completely removed old python project & source code
 
-[Unreleased]: https://github.com/recyclarr/recyclarr/compare/v8.7.0...HEAD
+[Unreleased]: https://github.com/recyclarr/recyclarr/compare/v8.7.1...HEAD
+[8.7.1]: https://github.com/recyclarr/recyclarr/compare/v8.7.0...v8.7.1
 [8.7.0]: https://github.com/recyclarr/recyclarr/compare/v8.6.0...v8.7.0
 [8.6.0]: https://github.com/recyclarr/recyclarr/compare/v8.5.1...v8.6.0
 [8.5.1]: https://github.com/recyclarr/recyclarr/compare/v8.5.0...v8.5.1